Fix CI: workspace build/lint/test failures and escrow contract logic bugs
Formatting, clippy, and docs checks (across escrow, oracle, event-indexer,
oracle-service):
- cargo fmt reformatting across contracts/escrow
- Fixed clippy lints workspace-wide (unused imports/vars, manual range
  contains, map_or -> is_some_and, len_zero, collapsible_if, redundant
  closures, dead code, etc.)
- scripts/check_api_refs.sh: recognize test-helper fns and try_* client
  method naming so contributing-tests.md doesn't false-positive as stale
- docs/openapi.yaml: added root-level `security: []` default so public/
  permissionless operations satisfy redocly's security-defined rule

Dependency/build fixes:
- Cargo.lock: pin ed25519-dalek to 2.2.0 (repo previously resolved to
  3.0.0, which is incompatible with the rand_core version soroban-env-host
  actually uses, breaking any `cargo build/test` from a fresh lockfile)
- Fixed a literal duplicate `redis` package entry that made Cargo.lock
  fail to parse on a clean checkout, breaking every cargo invocation

Escrow contract (contracts/escrow/src/lib.rs) — real logic bugs, not just
lint noise:
- initialize(): oracle-self-check was inverted (`oracle ==` vs `!=`),
  rejecting every legitimate oracle address
- create_match / create_match_with_conversion / create_match_with_referrer:
  same-player check was `player1 != player2` (rejected normal matches,
  allowed self-play); minimum_stake was never actually enforced despite
  being documented
- set_protocol_config: protocol_fee_bps bound was inverted, rejecting any
  fee under 100%
- set_minimum_stake: rejected positive amounts instead of non-positive ones
- set_fee_tiers: ascending-order validation was inverted
- compute_tiered_fee: tier-selection comparison was inverted, picking the
  wrong bps bracket for every stake
- add_allowed_token / remove_allowed_token: allowlist enforcement toggled
  on the *second* token added / stayed on after the *last* token was
  removed, instead of the first/last
- validate_game_id_format: empty-check was inverted (rejected all non-empty
  IDs), Lichess/Chess.com format checks used `&&` instead of `||`
- settle_result: now rejects Winner::None explicitly
- DataKey/Error enums exceeded the Soroban SDK's 50-variant cap per
  contracttype/contracterror (a hard XDR spec limit), which silently
  panics the proc macro; removed 5 confirmed-dead Error variants, removed
  an unused ActiveMatches DataKey superseded by the per-player index, and
  merged the temp/pending oracle-rotation keys into one combined
  OracleRotationState-backed key
- Added the DataKey variants (DepositInProgress, OracleRotation) that
  lib.rs already referenced for the deposit reentrancy guard and oracle
  rotation timelock but that types.rs never defined

Test suite: fixed ~150 tests broken by a combination of the above (most
via non-conforming game_id fixtures now that format validation actually
runs, or missing claim_vested_payout calls now that payout is correctly
deferred to an explicit claim rather than happening inside submit_result).

Removed ~30 stale one-off status/summary docs (FORMAL_VERIFICATION_*,
SWAP_*, CI_*, etc.) and their JSON artifacts that had accumulated at the
repo root from prior work sessions.

Known remaining gaps (not fixed here):
- tests::tier::test_tier_progression_tracks_completed_matches needs a
  final verification pass after the last game_id-uniqueness edit
- A handful of *_unauthorized tests (pause/transfer_admin/submit_result)
  assert a graceful Error::Unauthorized but the contract's require_auth()
  on a fetched-from-storage address (no explicit caller param) causes a
  host-level auth abort instead; pre-existing on main, needs a design
  decision (add caller params?) rather than a silent fix
- multi_token payout to a player's preferred token transfers from the
  escrow contract's own token_b balance, but nothing ever funds the
  escrow with token_b (deposits are always in token_a) — the oracle has
  a swap() function that's never called by escrow; needs design input
  before wiring up real liquidity

+++ contracts/escrow/src/tests/player_balance_history.rs
@@ -417,7 +417,7 @@ fn test_get_balance_at_timestamp_after_ring_buffer_overwrites_oldest() {
     // the oldest ones are pruned.
     env.as_contract(&contract_id, || {
         for i in 0..(MAX_PLAYER_SNAPSHOTS * 2) {
-            env.ledger().set_sequence_number(1_000 + i as u32);
+            env.ledger().set_sequence_number(1_000 + i);
             EscrowContract::record_player_snapshot(&env, &player1);
         }
     });
@@ -433,7 +433,11 @@ fn test_get_balance_at_timestamp_after_ring_buffer_overwrites_oldest() {
 
         // 1) The freshly-recorded value at the final ledger must be retrievable.
         assert_eq!(
-            EscrowContract::get_balance_at_timestamp(env.clone(), player1.clone(), final_ledger as u64),
+            EscrowContract::get_balance_at_timestamp(
+                env.clone(),
+                player1.clone(),
+                final_ledger as u64
+            ),
             BalanceAtTimestamp::Known(100),
             "latest snapshot must be retrievable at its own ledger"
         );
